refactor(models): retire configured-list adapters
