+++ src/commands/models/shared.ts
@@ -37,7 +37,12 @@ import {
 } from "./provider-aliases.js";
 
 export { formatTokenK } from "./list.format.js";
-export { ensureFlagCompatibility } from "./list.options.js";
+/** Rejects conflicting machine-readable output modes. */
+export function ensureFlagCompatibility(opts: { json?: boolean; plain?: boolean }): void {
+  if (opts.json && opts.plain) {
+    throw new Error("Choose either --json or --plain, not both.");
+  }
+}
 
 /** Formats millisecond durations for model command output. */
 export const formatMs = (value?: number | null) => {
